Renamed controller and test class
